telegram/requirements.txt
alpaca-py
requests